[refactoring] ssdlib h/cpp 분리
