+++ test-shell/main.cpp
@@ -1,4 +1,4 @@
-#include "ssdlib.cpp"
+#include "ssdlib.h"
 #include "shell.h"
 #include "command_factory.h"
