+++ test-shell/test-shell.vcxproj
@@ -185,6 +185,7 @@
     <ClInclude Include="ReadCommand.h" />
     <ClInclude Include="shell.h" />
     <ClInclude Include="ssd.h" />
+    <ClInclude Include="ssdlib.h" />
     <ClInclude Include="TestApp1Command.h" />
     <ClInclude Include="TestApp2Command.h" />
     <ClInclude Include="WriteCommand.h" />
